app2/guest360_docker/ingestion/google_pubsub/requirements.txt
boto3==1.26.59
google-cloud-pubsub==2.18.4
grpcio==1.58.0
grpcio-tools==1.58.0